+++ sdk/application/Catalog/CatalogApplicationValidator.d.ts
@@ -246,6 +246,7 @@ export = CatalogApplicationValidator;
  *   to retreive the nearest stores, e.g. 72.8691788
  * @property {number} [longitude] - Longitude of the location from where one
  *   wants to retreive the nearest stores, e.g. 19.1174114
+ * @property {string} [tags] - Search stores based on tags.
  */
 /**
  * @typedef UnfollowByIdParam
@@ -785,6 +786,10 @@ type GetStoresParam = {
      * wants to retreive the nearest stores, e.g. 19.1174114
      */
     longitude?: number;
+    /**
+     * - Search stores based on tags.
+     */
+    tags?: string;
 };
 type UnfollowByIdParam = {
     /**