+++ sdk/platform/Order/OrderPlatformValidator.js
@@ -219,6 +219,7 @@ const OrderPlatformModel = require("./OrderPlatformModel");
  * @property {number} [pageNo] - Page number for paginated data
  * @property {number} [pageSize] - Page size of data received per page
  * @property {boolean} [fetchActiveShipment] - Flag to fetch active shipments
+ * @property {boolean} [allowInactive] - Flag to allow inactive shipments
  * @property {boolean} [excludeLockedShipments] - Flag to fetch locked shipments
  * @property {string} [paymentMethods] - Comma separated values of payment methods
  * @property {string} [channelShipmentId] - App Shipment Id
@@ -595,6 +596,7 @@ class OrderPlatformValidator {
       pageNo: Joi.number(),
       pageSize: Joi.number(),
       fetchActiveShipment: Joi.boolean(),
+      allowInactive: Joi.boolean(),
       excludeLockedShipments: Joi.boolean(),
       paymentMethods: Joi.string().allow(""),
       channelShipmentId: Joi.string().allow(""),